libkz: move from module directory
Relocate libkz from src/modules to src/common.

This is built as part of libflux-internal.la and is not
part of the public API.  It was located in the modules
directory because of its dependency on the KVS API, which
is now built with libflux-core.

Various Makefile.am and include paths are updated.
