+++ configure.ac
@@ -260,6 +260,7 @@ AC_CONFIG_FILES( \
   src/common/libpmi/Makefile \
   src/common/libflux/Makefile \
   src/common/libkvs/Makefile \
+  src/common/libkz/Makefile \
   src/common/libsubprocess/Makefile \
   src/common/libcompat/Makefile \
   src/common/liboptparse/Makefile \
@@ -279,7 +280,6 @@ AC_CONFIG_FILES( \
   src/modules/connector-local/Makefile \
   src/modules/kvs/Makefile \
   src/modules/content-sqlite/Makefile \
-  src/modules/libkz/Makefile \
   src/modules/libjsc/Makefile \
   src/modules/barrier/Makefile \
   src/modules/wreck/Makefile \
