+++ src/common/libkz/kz.c
