+++ src/common/libkz/kz.h
